+++ bench/src/main/scala/geotrellis/spark/store/index/MergeQueueBench.scala
@@ -43,13 +43,13 @@ class MergeQueueBench {
   def setupData(): Unit = {
     ranges = Vector.empty[(BigInt, BigInt)]
 
-    for(i ← 0 until size) {
+    for (i ← 0 until size) {
       val start: BigInt = BigInt(skip.toLong * i)
       val end: BigInt = start + span
       ranges = ranges :+ (start, end)
     }
 
-    if(shuffle) {
+    if (shuffle) {
       ranges = Random.shuffle(ranges)
     }
   }
@@ -59,19 +59,24 @@ class MergeQueueBench {
 
   @Benchmark
   def insertAlternateRange(): Seq[(Long, Long)] =
-    MergeQueueBench.AlternateMergeQueue(ranges.map { case (a, b) => (a.longValue, b.longValue) })
+    MergeQueueBench.AlternateMergeQueue(ranges.map {
+      case (a, b) => (a.longValue, b.longValue)
+    })
 }
 
 object MergeQueueBench {
   object AlternateMergeQueue {
     def apply(ranges: TraversableOnce[(Long, Long)]): Seq[(Long, Long)] = {
       var merged: List[(Long, Long)] = Nil
-      ranges.toSeq.sortBy(_._1).foreach(r => {
-        merged = merged match {
-          case a :: rest if r._1 - 1 <= a._2 => (a._1, Math.max(a._2, r._2)) :: rest
-          case _ => r :: merged
-        }
-      })
+      ranges.toSeq
+        .sortBy(_._1)
+        .foreach(r => {
+          merged = merged match {
+            case a :: rest if r._1 - 1 <= a._2 =>
+              (a._1, Math.max(a._2, r._2)) :: rest
+            case _ => r :: merged
+          }
+        })
       merged.reverse
     }
   }