+++ spark/src/test/scala/geotrellis/spark/summary/polygonal/HistogramSpec.scala
@@ -18,23 +18,23 @@ package geotrellis.spark.summary.polygonal
 
 import geotrellis.spark._
 import geotrellis.spark.store.hadoop._
-import geotrellis.spark.testkit.testfiles._
+import geotrellis.spark.testkit.testfiles.TestFiles
 import geotrellis.raster._
 import geotrellis.raster.summary.polygonal._
+import geotrellis.raster.summary.polygonal.visitors._
 import geotrellis.vector._
 import geotrellis.spark.testkit._
-
-import org.apache.spark.SparkContext
 import org.scalatest.FunSpec
 
 import collection.immutable.HashMap
 
 class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
-
   describe("Histogram Zonal Summary Operation") {
+
     val modHundred = Mod10000TestFile
     val multiModHundred = modHundred.withContext { _.mapValues { MultibandTile(_) }}
     val ones = AllOnesTestFile
+    val inc = IncreasingTestFile
     val multi = ones.withContext { _.mapValues { tile => MultibandTile(tile, tile, tile) }}
 
     val tileLayout = modHundred.metadata.tileLayout
@@ -72,44 +72,62 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
     }
 
     it("should get correct histogram over whole raster extent") {
-      val histogram = modHundred.polygonalHistogram(totalExtent.toPolygon)
+      // We use FastMapHistogram for this test because StreamingHistogram can shuffle bucket
+      //  bounds and counts based on the order in which dissimilar elements are added. This addition is
+      //  non-deterministic for our RDD polygonal summaries.
+      val histogram = modHundred.polygonalSummaryValue(totalExtent.toPolygon, FastMapHistogramVisitor).toOption.get
+      val expected = modHundred.stitch.polygonalSummary(totalExtent.toPolygon, FastMapHistogramVisitor).toOption.get
+
+      histogram.totalCount should be (expected.totalCount)
+      histogram.foreachValue(v => histogram.itemCount(v) should be (expected.itemCount(v)))
 
       var map = HashMap[Int, Int]()
-
       for (i <- 0 until count) {
         val key = i % 10000
         val v = map.getOrElse(key, 0) + 1
         map = map + (key -> v)
       }
-
       map.foreach { case (v, k) => histogram.itemCount(v) should be (k) }
     }
 
+    // We use FastMapHistogram for this test because StreamingHistogram can shuffle bucket
+    //  bounds and counts based on the order in which dissimilar elements are added. This addition is
+    //  non-deterministic for our RDD polygonal summaries.
     it("should get correct histogram over whole raster extent for a MultibandTileRDD") {
-      val histogram = multiModHundred.polygonalHistogram(totalExtent.toPolygon).head
+      val histogram = multiModHundred.polygonalSummaryValue(
+        totalExtent.toPolygon,
+        FastMapHistogramVisitor
+      ).toOption.get.head
+      val expected = multiModHundred.stitch.polygonalSummary(
+        totalExtent.toPolygon,
+        FastMapHistogramVisitor
+      ).toOption.get.head
+
+      histogram.totalCount should be (expected.totalCount)
+      histogram.foreachValue(v => histogram.itemCount(v) should be (expected.itemCount(v)))
 
       var map = HashMap[Int, Int]()
-
       for (i <- 0 until count) {
         val key = i % 10000
         val v = map.getOrElse(key, 0) + 1
         map = map + (key -> v)
       }
-
       map.foreach { case (v, k) => histogram.itemCount(v) should be (k) }
     }
 
     it("should get correct histogram over a quarter of the extent") {
-      val histogram = ones.polygonalHistogram(quarterExtent.toPolygon)
-      val expected = ones.stitch.tile.polygonalHistogram(totalExtent, quarterExtent.toPolygon)
+      val histogram = inc.polygonalSummaryValue(
+        quarterExtent.toPolygon,
+        StreamingHistogramVisitor).toOption.get
+      val expected = inc.stitch.polygonalSummary(quarterExtent.toPolygon, StreamingHistogramVisitor).toOption.get
 
       histogram.minMaxValues should be (expected.minMaxValues)
       histogram.itemCount(1) should be (expected.itemCount(1))
     }
 
     it("should get correct histogram over a quarter of the extent for a MultibandTileRDD") {
-      val histogram = multi.polygonalHistogram(quarterExtent.toPolygon)
-      val expected = multi.stitch.tile.polygonalHistogram(totalExtent, quarterExtent.toPolygon)
+      val histogram = multi.polygonalSummaryValue(quarterExtent.toPolygon, StreamingHistogramVisitor).toOption.get
+      val expected = multi.stitch.polygonalSummary(quarterExtent.toPolygon, StreamingHistogramVisitor).toOption.get
 
       histogram.size should be (expected.size)
 
@@ -120,16 +138,16 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
     }
 
     it("should get correct histogram over half of the extent in diamond shape") {
-      val histogram = ones.polygonalHistogram(diamondPoly)
-      val expected = ones.stitch.tile.polygonalHistogram(totalExtent, diamondPoly)
+      val histogram = ones.polygonalSummaryValue(diamondPoly, StreamingHistogramVisitor).toOption.get
+      val expected = ones.stitch.polygonalSummary(diamondPoly, StreamingHistogramVisitor).toOption.get
 
       histogram.minMaxValues should be (expected.minMaxValues)
       histogram.itemCount(1) should be (expected.itemCount(1))
     }
 
     it("should get correct histogram over half of the extent in diamond shape for a MultibandTileRDD") {
-      val histogram = multi.polygonalHistogram(diamondPoly)
-      val expected = multi.stitch.tile.polygonalHistogram(totalExtent, diamondPoly)
+      val histogram = multi.polygonalSummaryValue(diamondPoly, StreamingHistogramVisitor).toOption.get
+      val expected = multi.stitch.polygonalSummary(diamondPoly, StreamingHistogramVisitor).toOption.get
 
       histogram.size should be (expected.size)
 
@@ -140,16 +158,16 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
     }
 
     it("should get correct histogram over polygon with hole") {
-      val histogram = ones.polygonalHistogram(polyWithHole)
-      val expected = ones.stitch.tile.polygonalHistogram(totalExtent, polyWithHole)
+      val histogram = ones.polygonalSummaryValue(polyWithHole, StreamingHistogramVisitor).toOption.get
+      val expected = ones.stitch.polygonalSummary(polyWithHole, StreamingHistogramVisitor).toOption.get
 
       histogram.minMaxValues should be (expected.minMaxValues)
       histogram.itemCount(1) should be (expected.itemCount(1))
     }
 
     it("should get correct histogram over polygon with hole for a MultibandTileRDD") {
-      val histogram = multi.polygonalHistogram(polyWithHole)
-      val expected = multi.stitch.tile.polygonalHistogram(totalExtent, polyWithHole)
+      val histogram = multi.polygonalSummaryValue(polyWithHole, StreamingHistogramVisitor).toOption.get
+      val expected = multi.stitch.polygonalSummary(polyWithHole, StreamingHistogramVisitor).toOption.get
 
       histogram.size should be (expected.size)
 
@@ -161,17 +179,17 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
   }
 
   describe("Histogram Zonal Summary Operation (collections api)") {
-    val modHundred = Mod10000TestFile.toCollection
+    val modHundred = Mod10000TestFile
     val multiModHundred = Mod10000TestFile.withContext { _.mapValues { MultibandTile(_) }}
-    val ones = AllOnesTestFile.toCollection
+    val ones = AllOnesTestFile
     val multi = AllOnesTestFile.withContext { _.mapValues { tile => MultibandTile(tile, tile, tile) }}
 
     val tileLayout = modHundred.metadata.tileLayout
-    val count = modHundred.length * tileLayout.tileCols * tileLayout.tileRows
+    val count = modHundred.toCollection.length * tileLayout.tileCols * tileLayout.tileRows
     val totalExtent = modHundred.metadata.extent
 
     it("should get correct histogram over whole raster extent") {
-      val histogram = modHundred.polygonalHistogram(totalExtent.toPolygon)
+      val histogram = modHundred.polygonalSummaryValue(totalExtent.toPolygon, FastMapHistogramVisitor).toOption.get
 
       var map = HashMap[Int, Int]()
 
@@ -185,7 +203,7 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
     }
 
     it("should get correct histogram over whole raster extent for MultibandTiles") {
-      val histogram = multiModHundred.polygonalHistogram(totalExtent.toPolygon).head
+      val histogram = multiModHundred.polygonalSummaryValue(totalExtent.toPolygon, FastMapHistogramVisitor).toOption.get.head
 
       var map = HashMap[Int, Int]()
 
@@ -209,8 +227,8 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
         totalExtent.ymin + yd / 2
       )
 
-      val histogram = ones.polygonalHistogram(quarterExtent.toPolygon)
-      val expected = ones.stitch.tile.polygonalHistogram(totalExtent, quarterExtent.toPolygon)
+      val histogram = ones.polygonalSummaryValue(quarterExtent.toPolygon, StreamingHistogramVisitor).toOption.get
+      val expected = ones.stitch.polygonalSummary(quarterExtent.toPolygon, StreamingHistogramVisitor).toOption.get
 
       histogram.minMaxValues should be (expected.minMaxValues)
       histogram.itemCount(1) should be (expected.itemCount(1))
@@ -227,8 +245,8 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
         totalExtent.ymin + yd / 2
       )
 
-      val histogram = multi.polygonalHistogram(quarterExtent.toPolygon)
-      val expected = multi.stitch.tile.polygonalHistogram(totalExtent, quarterExtent.toPolygon)
+      val histogram = multi.polygonalSummaryValue(quarterExtent.toPolygon, StreamingHistogramVisitor).toOption.get
+      val expected = multi.stitch.polygonalSummary(quarterExtent.toPolygon, StreamingHistogramVisitor).toOption.get
 
       histogram.size should be (expected.size)
 
@@ -250,8 +268,8 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
 
       val poly = Polygon(Line(Array(p1, p2, p3, p4, p1)))
 
-      val histogram = ones.polygonalHistogram(poly)
-      val expected = ones.stitch.tile.polygonalHistogram(totalExtent, poly)
+      val histogram = ones.polygonalSummaryValue(poly, StreamingHistogramVisitor).toOption.get
+      val expected = ones.stitch.polygonalSummary(poly, StreamingHistogramVisitor).toOption.get
 
       histogram.minMaxValues should be (expected.minMaxValues)
       histogram.itemCount(1) should be (expected.itemCount(1))
@@ -268,8 +286,8 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
 
       val poly = Polygon(Line(Array(p1, p2, p3, p4, p1)))
 
-      val histogram = multi.polygonalHistogram(poly)
-      val expected = multi.stitch.tile.polygonalHistogram(totalExtent, poly)
+      val histogram = multi.polygonalSummaryValue(poly, StreamingHistogramVisitor).toOption.get
+      val expected = multi.stitch.polygonalSummary(poly, StreamingHistogramVisitor).toOption.get
 
       histogram.size should be (expected.size)
 
@@ -301,8 +319,8 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
       val interior = Line(Array(pi1, pi2, pi3, pi4, pi1))
       val poly = Polygon(exterior, interior)
 
-      val histogram = ones.polygonalHistogram(poly)
-      val expected = ones.stitch.tile.polygonalHistogram(totalExtent, poly)
+      val histogram = ones.polygonalSummaryValue(poly, StreamingHistogramVisitor).toOption.get
+      val expected = ones.stitch.polygonalSummary(poly, StreamingHistogramVisitor).toOption.get
 
       histogram.minMaxValues should be (expected.minMaxValues)
       histogram.itemCount(1) should be (expected.itemCount(1))
@@ -330,8 +348,8 @@ class HistogramSpec extends FunSpec with TestEnvironment with TestFiles {
       val interior = Line(Array(pi1, pi2, pi3, pi4, pi1))
       val poly = Polygon(exterior, interior)
 
-      val histogram = multi.polygonalHistogram(poly)
-      val expected = multi.stitch.tile.polygonalHistogram(totalExtent, poly)
+      val histogram = multi.polygonalSummaryValue(poly, StreamingHistogramVisitor).toOption.get
+      val expected = multi.stitch.polygonalSummary(poly, StreamingHistogramVisitor).toOption.get
 
       histogram.size should be (expected.size)